Move TreeStatus from editable-tree to editable-tree-2
