+++ experimental/dds/tree2/src/feature-libraries/editable-tree-2/index.ts
@@ -22,6 +22,7 @@ export {
 	TypedNodeUnion,
 	boxedIterator,
 	CheckTypesOverlap,
+	TreeStatus,
 } from "./editableTreeTypes";
 
 export {
