+++ experimental/dds/tree2/src/feature-libraries/editable-tree-2/lazyEntity.ts
@@ -9,10 +9,9 @@ import {
 	ITreeSubscriptionCursor,
 	ITreeSubscriptionCursorState,
 } from "../../core";
-import { TreeStatus } from "../editable-tree";
 import { fail, disposeSymbol, IDisposable } from "../../util";
 import { Context } from "./context";
-import { Tree, boxedIterator } from "./editableTreeTypes";
+import { Tree, TreeStatus, boxedIterator } from "./editableTreeTypes";
 
 /**
  * Declare an enumerable own property on `T` under the key `key` using the implementation of one on `from`.
