+++ experimental/dds/tree2/src/feature-libraries/editable-tree-2/lazyTree.ts
@@ -35,7 +35,7 @@ import {
 	StructSchema,
 	Any,
 } from "../typed-schema";
-import { TreeStatus, treeStatusFromPath } from "../editable-tree";
+import { treeStatusFromPath } from "../editable-tree";
 import { EditableTreeEvents } from "../untypedTree";
 import { FieldKinds } from "../default-field-kinds";
 import { Context } from "./context";
@@ -51,6 +51,7 @@ import {
 	TreeField,
 	TreeNode,
 	boxedIterator,
+	TreeStatus,
 } from "./editableTreeTypes";
 import { makeField } from "./lazyField";
 import {
